requirements.txt
streamlit==1.31.0
boto3==1.34.40
Pillow>=10.0.0
